refactor: restructure and clean up code. Fix error pop-up issue
