+++ babel.config.js
@@ -2,10 +2,10 @@ module.exports = {
   presets: ['module:metro-react-native-babel-preset'],
   plugins: [
     [
-      'react-native-reanimated/plugin',
-      {
+      { 
         globals: ['__scanPoseLandmarks'],
       },
+      'react-native-reanimated/plugin',
     ],
   ],
 };
