+++ example/package.json
@@ -10,8 +10,7 @@
     "pods": "pod-install --quiet",
     "build:APK": "react-native bundle --platform android --dev false --entry-file index.tsx --bundle-output android/app/src/main/assets/index.android.bundle --assets-dest android/app/src/main/res",
     "create:debugAPK": "cd android && ./gradlew assembleDebug",
-    "postinstall": "patch-package && node scripts/fix-duplicate-resource/android-release-gradle-fix.js" 
-
+    "postinstall": "patch-package && node scripts/fix-duplicate-resource/android-release-gradle-fix.js"
   },
   "dependencies": {
     "@react-navigation/native": "^6.0.13",
@@ -23,8 +22,8 @@
     "react-native-safe-area-context": "^4.4.1",
     "react-native-screens": "^3.18.2",
     "react-native-simple-radio-button": "^2.7.4",
-    "react-native-vision-camera": "2.13.0",
-    "react-native-svg": "12.3.0"
+    "react-native-svg": "12.3.0",
+    "react-native-vision-camera": "2.13.0"
   },
   "devDependencies": {
     "@babel/core": "^7.12.10",
@@ -37,6 +36,6 @@
     "metro-react-native-babel-preset": "^0.67.0",
     "patch-package": "^6.4.7",
     "postinstall-postinstall": "^2.1.0",
-    "react-native-flipper": "^0.174.0"
+    "react-native-flipper": "^0.179.0"
   }
 }
