+++ src/components/assessment/index.tsx
@@ -1,0 +1,2 @@
+export * from './interface'
+export * from './component'
