+++ src/components/index.tsx
@@ -1,1 +1,1 @@
-export * from './Assessments'
+export * from './assessment'
